Support for Client string in SOAP requests. This is for Salesforce partners who are receiving the "API_DISABLED_FOR_ORG" error when trying to connect to Salesforce professional accounts.
To take advantage of this, you must add the client name to the OAuth parameters when logging in.
Example..

// Previously
OAuthConnectorConfig parameters = new OAuthConnectorConfig(CLIENT_ID, REDIRECT_URI);
// This is new
parameters.setClient(CLIENT_STRING);
// Previously
Salesforce.loginOAuth(this, parameters, new LoginResponseListener());

This also means that when updating an existing app, make sure that the user logs out and back in.